Made player better + weapon pickup
The player is now a reasonable size and I have a quick and dirty solution for the weapon pickup. It's awful but essentially we'll have an invisible version of the shotgun where the pickup is and when the player touches the pickup the invisible shotgun will be added to the player and stuff.

I know it's bad sue me.